+++ src/demo/blinkdemo.cpp
@@ -2,6 +2,7 @@
  * BlinkDemo - Displays a simple blink animation.
  */
 
+#include <memory>
 #include "blinkdemo.h"
 #include "colorpresets.h"
 #include "controller/maestrocontroller.h"
@@ -12,7 +13,8 @@ namespace PixelMaestroStudio {
 	BlinkDemo::BlinkDemo(QWidget* parent, MaestroController* maestro_controller) : MaestroDrawingArea(parent, maestro_controller) {
 		Section* section = maestro_controller_->set_sections(1, Point(10, 10));
 
-		Animation* animation = section->set_animation(AnimationType::Blink, ColorPresets::Colorwheel, 12);
+		Animation* animation = section->set_animation(AnimationType::Blink);
+		animation->set_palette(new Palette(ColorPresets::Colorwheel, 12, false));
 		animation->set_timer(1000);
 	}
 }